requirements.txt
attrs==23.2.0
azure-ai-formrecognizer==3.3.3
azure-ai-ml==1.16.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0
azure-storage-file-share==12.16.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
google-api-core==2.19.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
idna==3.7
isodate==0.6.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
marshmallow==3.21.2
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
packaging==24.0
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydash==8.0.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
six==1.16.0
strictyaml==1.7.3
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
